+++ test/forwarder/Forwarder.t.sol
@@ -722,7 +722,7 @@ contract ForwarderTest is Test {
             to: to,
             paymentToken: _paymentToken,
             data: data,
-            deadline: block.timestamp + 30 days,
+            deadline: uint64(block.timestamp + 30 days),
             nonce: nonce
         });
